refactor: extract mail transport port
