+++ CONTRACT.md
@@ -4,6 +4,7 @@ contract_version: 2
 related_files:
   - ANATOMY.md
   - src/lingtai/kernel/event_journal/CONTRACT.md
+  - src/lingtai/kernel/mail_transport/CONTRACT.md
   - CONTRIBUTING.md
   - README.md
   - dev-guide-skill/SKILL.md
