+++ src/lingtai/adapters/posix/__init__.py
@@ -1,5 +1,6 @@
 """Narrow POSIX filesystem adapters."""
 
 from .event_journal import PosixJsonlEventJournalAdapter
+from .mail import PosixFilesystemMailAdapter
 
-__all__ = ["PosixJsonlEventJournalAdapter"]
+__all__ = ["PosixJsonlEventJournalAdapter", "PosixFilesystemMailAdapter"]
