+++ src/lingtai/kernel/ANATOMY.md
@@ -22,6 +22,7 @@ related_files:
   - src/lingtai/kernel/config_resolve.py
   - src/lingtai/kernel/event_journal/ANATOMY.md
   - src/lingtai/kernel/handshake.py
+  - src/lingtai/kernel/mail_transport/ANATOMY.md
   - src/lingtai/kernel/i18n/ANATOMY.md
   - src/lingtai/tools/ANATOMY.md
   - src/lingtai/tools/notification/__init__.py
@@ -86,6 +87,7 @@ The kernel root holds the coordinator (`base_agent/`) plus a flat collection of 
 - `maintenance/` — explicit, operator-run maintenance reporters. `retention.py` implements the dry-run-only retention report for stale terminal daemon run dirs, old sent-mail copies, opt-in archive mail, rebuildable `logs/log.sqlite`, and read-only high-footprint observations for portal replay, agent logs, and history; it never deletes or rewrites files. See `maintenance/ANATOMY.md`.
 - `nudge/` — **per-agent periodic checks** that publish notification reminders when something needs the agent's attention. `__init__.py:run_checks(agent)` is called once per heartbeat tick from `base_agent/lifecycle.py:_heartbeat_loop` (~line 320, wrapped in try/except). Each check is a self-contained module under `nudge/` exposing `check(agent) -> None`. `kernel_version.py` uses the shared `.notification/nudge.json` multi-entry payload for packaged runtime refresh/package-update reminders; `source_drift.py` compares startup and on-disk runtime fingerprints for non-dev runtimes and skips editable/source/dev checkouts; `goal.py` is an IDLE-only check that reads protected `.notification/goal.json`; if the file exists, is active, and the idle delay has elapsed, it publishes one short `goal.reminder` event into `.notification/system.json` pointing back to `goal.json` and the goal manual. See `nudge/ANATOMY.md` for details and the "add a new nudge" recipe.
 - `event_journal/` — Core-owned outbound Port for append-only structured events plus authoritative `JournalPosition` provenance. `BaseAgent` receives this Port; POSIX JSONL/SQLite composition lives outside Core. See `event_journal/ANATOMY.md` (`event_journal/__init__.py:9-26`).
+- `mail_transport/` — Core-owned outbound Port for fire-and-forget agent messaging (`MailTransportPort`: `send`/`listen`/`stop`/`address`). `BaseAgent` receives this Port as `mail_service`; the concrete POSIX filesystem mailbox (`PosixFilesystemMailAdapter`) lives outside Core and is wired by the CLI. See `mail_transport/ANATOMY.md` and `mail_transport/CONTRACT.md` (`mail_transport/__init__.py:16-62`).
 - `runtime_identity.py` — compact kernel runtime identity helper for durable event logs. `runtime_identity_event_fields()` (`runtime_identity.py:18-27`) returns the top-level `kernel_version` / `kernel_runtime_stamp` fields plus the nested `kernel_runtime` payload; `runtime_identity()` (`runtime_identity.py:30-80`) computes a cached release/dev stamp from package metadata, editable/source checkout detection, and best-effort git state without importing the wrapper `lingtai` package.
 - `_fsutil.py` — shared, stdlib-only filesystem / JSON / JSONL foundation for kernel state writes. It centralizes sibling-temp atomic replacement, JSON encoding defaults, `read_json(default=...)`, append-only JSONL offsets, and UTC timestamp formatting (`_fsutil.py:1-26`, `_fsutil.py:76-271`).
 - `notifications.py` — **canonical `.notification/` filesystem helpers**. Channel names are syntax-checked and then allowlisted: built-ins (`email`, `soul`, `system`, `goal`, `nudge`, `molt`, `post-molt`, `bash`, `cron`, `btw`, `tool_loop_guard`) plus dynamic `mcp.*` bridge channels. Unknown files in `.notification/` are ignored by `notification_fingerprint()` and `collect_notifications()` and rejected by helper publish/clear/dismiss calls. `notification_fingerprint()` is byte-content-based (`name`, byte count, SHA-256) so byte-identical producer rewrites do not wake/reinject solely because mtime changed. `publish(workdir, tool_name, payload)` writes one file atomically (tmp+rename); `clear(workdir, tool_name)` deletes one file (idempotent); `submit(workdir, tool_name, *, data, header, icon, priority, instructions=None)` is the canonical producer-facing helper. `dismiss_channel()` is the strict agent-facing generic dismiss path: protected source-of-truth channels such as `goal` refuse generic dismiss, while `system` additionally supports atomic single-event removal by `event_id` or `ref_id` without deleting the rest of `system.json`. The `system` intrinsic re-exports `submit` as `publish_notification` and `clear` as `clear_notification`.
