+++ src/lingtai/services/mail.py
@@ -1,4 +1,19 @@
-"""Re-export kernel mail services."""
-from lingtai.kernel.services.mail import MailService, FilesystemMailService
+"""Re-export the mail transport Port and its production POSIX adapter.
 
-__all__ = ["MailService", "FilesystemMailService"]
+This wrapper is a high-level composition surface, so it may import the adapter.
+``MailService`` is the Core-owned Port; ``FilesystemMailService`` is the
+back-compat public name for the production adapter
+(``PosixFilesystemMailAdapter``). Both names resolve to a single implementation
+each — no shim, no dual implementation.
+"""
+from lingtai.kernel.mail_transport import MailTransportPort as MailService
+from lingtai.adapters.posix.mail import (
+    PosixFilesystemMailAdapter,
+    PosixFilesystemMailAdapter as FilesystemMailService,
+)
+
+__all__ = [
+    "MailService",
+    "FilesystemMailService",
+    "PosixFilesystemMailAdapter",
+]
