+++ src/lingtai/tools/email/ANATOMY.md
@@ -50,7 +50,7 @@ Filesystem-based email system — mailbox I/O, composition, search, contacts, re
 ## Connections
 
 - **Inbound:** `handle()` is called by the tool dispatcher (via `base_agent._dispatch_tool`). `boot()` is called during agent construction in `base_agent/__init__.py`.
-- **Inbound (cross-module):** `_new_mailbox_id` is now owned by the kernel mail service — defined at `src/lingtai/kernel/services/mail.py:29` and used there by `send()` (`src/lingtai/kernel/services/mail.py:181`). The email package imports and re-exports it via `primitives.py:20` for back-compat with `lingtai.tools.email._new_mailbox_id` importers.
+- **Inbound (cross-module):** `_new_mailbox_id` is owned by the kernel mail service — defined at `src/lingtai/kernel/services/mail.py:29-44`. After the mail Ports & Adapters split it is imported explicitly by the POSIX mail adapter (`src/lingtai/adapters/posix/mail.py:29`) and consumed by its `send()` (`src/lingtai/adapters/posix/mail.py:118`), not by `kernel/services/mail.py` (which no longer defines a transport). The email package imports and re-exports it via `primitives.py:20` for back-compat with `lingtai.tools.email._new_mailbox_id` importers.
 - **Inbound (cross-module):** `EmailManager` is imported by `src/lingtai/__init__.py:19` for the wrapper re-export.
 - **Outbound:** Depends on `..i18n` (translations), `..message` (message construction), `..time_veil` (timestamp scrubbing), `..token_counter` (budget checks in `_check`).
 - **Outbound (unread-email producer):** Mail arrival writes `.notification/email.json` via `publish_notification` (or deletes it via `clear_notification` when count hits 0). `base_agent/messaging.py:_on_normal_mail` calls `_rerender_unread_digest(agent)` (resolved via `_intrinsic_hook("email", "_rerender_unread_digest")` at `src/lingtai/kernel/base_agent/messaging.py:61`) which uses `primitives.py:_render_unread_digest` for count/newest compatibility and `_unread_notification_context` for full-body entries, then `system.publish_notification(workdir, "email", header=…, icon="📧", data={count, newest_received_at, email_ids, emails})`. The kernel's `_sync_notifications` poll picks up the fingerprint change on the next heartbeat tick and updates the wire's `notification(action="check")` block. See root `ANATOMY.md` "Notifications" for the full architecture.
